externalize css and ts from vue for better ide support
